Refactor Add/Edit Account forms to use custom hooks and separated logic
- Extracted form state and database logic into `useAddAccountForm` and `useEditAccountForm` hooks.
- Extracted purely logical UI visibility calculations (`canShowAER`, `canShowBonus`) into `accountRules.ts`.
- Cleaned up `AddAccountPage.tsx` and `EditAccountPage.tsx` to strictly be presentational components invoking these new abstractions.
- Dynamic string handling for "Bonus Rate" vs "Maturity Tracking" maintained directly in JSX.
