+++ src/pages/EditAccountPage.tsx
@@ -1,119 +1,47 @@
-import { useState, useEffect } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
-import { useLiveQuery } from 'dexie-react-hooks';
-import { db } from '@/lib/db';
-import { useStore } from '@/store/useStore';
 import { Icon } from '@/components/ui/Icon';
 import { ACCOUNT_CATEGORIES, type AccountCategory } from '@/constants/taxConstants';
 import { InterestLedger } from '@/components/accounts/InterestLedger';
+import { useEditAccountForm } from '@/hooks/useEditAccountForm';
 
 export function EditAccountPage() {
-    const { id } = useParams<{ id: string }>();
-    const navigate = useNavigate();
-    const { profile } = useStore();
-    const p1Name = profile?.partner1Name || 'Partner 1';
-    const p2Name = profile?.partner2Name || 'Partner 2';
-
-    const account = useLiveQuery(() => db.accounts.get(id || ''), [id]);
-
-    const [accountName, setAccountName] = useState('');
-    const [nickname, setNickname] = useState('');
-    const [last4Digits, setLast4Digits] = useState('');
-    const [category, setCategory] = useState<AccountCategory | ''>('');
-    const [balance, setBalance] = useState('');
-    const [interestRate, setInterestRate] = useState('');
-    const [budgetOrder, setBudgetOrder] = useState('');
-    const [interestTrackingMethod, setInterestTrackingMethod] = useState<'aer' | 'manual'>('aer');
-    const [bonusRateActive, setBonusRateActive] = useState(false);
-    const [bonusEndDate, setBonusEndDate] = useState('');
-    const [payoutFrequency, setPayoutFrequency] = useState<'monthly' | 'annually' | 'at_maturity' | ''>('');
-    const [payoutDate, setPayoutDate] = useState('');
-    const [ownerId, setOwnerId] = useState('joint');
-
-    const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
-
-    useEffect(() => {
-        if (account) {
-            setAccountName(account.name);
-            setNickname(account.nickname || '');
-            setLast4Digits(account.last4Digits || '');
-            setCategory((account.category as AccountCategory) || '');
-            setBalance(account.balance.toString());
-            setInterestRate(account.interestRate.toString());
-            setBudgetOrder(account.budgetOrder !== undefined ? account.budgetOrder.toString() : '');
-            setInterestTrackingMethod(account.interestTrackingMethod || 'aer');
-            setBonusRateActive(account.bonusRateActive || false);
-            if (account.bonusEndDate) {
-                // Convert timestamp to YYYY-MM-DD
-                const date = new Date(account.bonusEndDate);
-                const year = date.getFullYear();
-                const month = String(date.getMonth() + 1).padStart(2, '0');
-                const day = String(date.getDate()).padStart(2, '0');
-                setBonusEndDate(`${year}-${month}-${day}`);
-            }
-            setPayoutFrequency(account.interestPayoutFrequency || '');
-            if (account.interestPayoutDate) {
-                const pDate = new Date(account.interestPayoutDate);
-                const pYear = pDate.getFullYear();
-                const pMonth = String(pDate.getMonth() + 1).padStart(2, '0');
-                const pDay = String(pDate.getDate()).padStart(2, '0');
-                setPayoutDate(`${pYear}-${pMonth}-${pDay}`);
-            } else {
-                setPayoutDate('');
-            }
-            setOwnerId(account.ownerId);
-        }
-    }, [account]);
-
-    const isEligibleForAER = category === '' || !['Stocks & Shares', 'Shares ISA', 'DC Pension', 'Premium Bonds'].includes(category as string);
-    const showBonus = category === '' || ['Easy Access Savings', 'Cash ISA', 'Current Account', 'Fixed Term Savings'].includes(category as string);
-
-    useEffect(() => {
-        if (!isEligibleForAER && interestTrackingMethod !== 'manual') {
-            setInterestTrackingMethod('manual');
-        }
-    }, [isEligibleForAER, interestTrackingMethod]);
-
-    const handleSave = async () => {
-        if (!id || !accountName || !balance || !category) return;
-        if (isEligibleForAER && interestTrackingMethod === 'aer' && !interestRate) return;
-
-        const finalInterestRate = isEligibleForAER && interestTrackingMethod === 'aer' ? parseFloat(interestRate) : 0;
-
-        try {
-            await db.accounts.update(id, {
-                name: accountName,
-                nickname: nickname || undefined,
-                last4Digits: last4Digits || undefined,
-                category,
-                balance: parseFloat(balance),
-                interestRate: finalInterestRate,
-                interestTrackingMethod: isEligibleForAER ? interestTrackingMethod : 'manual',
-                budgetOrder: budgetOrder !== '' ? parseInt(budgetOrder, 10) : undefined,
-                bonusRateActive: showBonus ? bonusRateActive : false,
-                bonusEndDate: showBonus && bonusRateActive && bonusEndDate ? new Date(bonusEndDate).getTime() : undefined,
-                interestPayoutFrequency: payoutFrequency || undefined,
-                interestPayoutDate: (payoutFrequency === 'annually' || payoutFrequency === 'at_maturity') && payoutDate
-                    ? new Date(payoutDate).getTime()
-                    : undefined,
-                ownerId,
-                updatedAt: Date.now(),
-            });
-            navigate('/accounts');
-        } catch (error) {
-            console.error('Failed to update account', error);
-        }
-    };
-
-    const handleDelete = async () => {
-        if (!id) return;
-        try {
-            await db.accounts.delete(id);
-            navigate('/accounts');
-        } catch (error) {
-            console.error('Failed to delete account', error);
-        }
-    };
+    const {
+        account,
+        navigate,
+        p1Name,
+        p2Name,
+        accountName,
+        setAccountName,
+        nickname,
+        setNickname,
+        last4Digits,
+        setLast4Digits,
+        category,
+        setCategory,
+        balance,
+        setBalance,
+        interestRate,
+        setInterestRate,
+        budgetOrder,
+        setBudgetOrder,
+        interestTrackingMethod,
+        setInterestTrackingMethod,
+        bonusRateActive,
+        setBonusRateActive,
+        bonusEndDate,
+        setBonusEndDate,
+        payoutFrequency,
+        setPayoutFrequency,
+        payoutDate,
+        setPayoutDate,
+        ownerId,
+        setOwnerId,
+        showDeleteConfirm,
+        setShowDeleteConfirm,
+        isEligibleForAER,
+        showBonus,
+        handleSave,
+        handleDelete
+    } = useEditAccountForm();
 
     if (!account) return <div className="p-4 text-center">Loading...</div>;
 
@@ -307,7 +235,9 @@ export function EditAccountPage() {
                     <div className="p-5 rounded-xl border border-primary/20 bg-primary/5 space-y-4">
                         <div className="flex items-center justify-between">
                             <div>
-                                <h3 className="font-bold text-slate-900 dark:text-slate-100">Bonus Rate</h3>
+                                <h3 className="font-bold text-slate-900 dark:text-slate-100">
+                                    {category === 'Fixed Term Savings' || category === 'Notice Savings' ? 'Maturity Tracking' : 'Bonus Rate'}
+                                </h3>
                                 <p className="text-xs text-slate-500 dark:text-slate-400">Additional interest for a limited period</p>
                             </div>
                             <label className="relative inline-flex items-center cursor-pointer">
